move crates into a dedicated directory and start extracting types
