+++ crates/nym-bridge/Cargo.toml
@@ -1,0 +1,91 @@
+[package]
+name = "nym-bridge"
+version = "0.2.0"
+edition = "2024"
+publish = false
+authors = ["Nym Technologies SA"]
+description = "Server-side transparent forwarder that unwraps obfuscated Nym VPN bridge traffic."
+license = "AGPL-3.0-only"
+repository = "https://github.com/nymtech/nym-bridges"
+homepage = "https://nymtech.net"
+
+[dependencies]
+anyhow.workspace = true
+clap = { workspace = true, features = ["derive"] }
+quinn.workspace = true
+rustls = { workspace = true, default-features = false, features = [
+	"std",
+	"ring",
+	"logging",
+] }
+tokio = { workspace = true, features = ["full"] }
+tokio-rustls = { workspace = true, default-features = false, features = [
+	"ring",
+	"tls12",
+	"logging",
+] }
+tokio-util = { workspace = true, features = ["full"] }
+tracing.workspace = true
+tracing-subscriber = { workspace = true, default-features = false, features = [
+	"env-filter",
+	"fmt",
+	"ansi",
+	"time",
+	"local-time",
+] }
+
+
+nym-bridges = { path = "../nym-bridges" }
+
+nym-bin-common = "1.21.4"
+
+# ------------------------------------------------------------------------------
+# START DEBIAN PACKAGING
+#
+# Configurations for the cargo-deb cargo plugin which builds Debian packages in
+# target/debian/ when invoked with: cargo deb.
+#
+# Note that as the postinst script uses the adduser command we declare a
+# dependency on the adduser package to keep the lintian tool happy.
+#
+# See:
+#   - https://github.com/mmstick/cargo-deb
+#   - https://lintian.debian.org/tags/systemd-service-file-outside-lib.html
+#   - https://www.debian.org/doc/debian-policy/ch-files.html#behavior
+
+[package.metadata.deb]
+name = "nym-bridge"
+maintainer = "Nym Technologies SA <contact@nymtech.net>"
+priority = "optional"
+section = "net"
+license-file = ["../../LICENSE", "0"]
+extended-description-file = "pkg/doc/description.txt"
+depends = "$auto, adduser"
+maintainer-scripts = "pkg/debian/"
+copyright = "Copyright (c) 2025, Nym Technologies SA. All rights reserved."
+assets = [
+    [
+        "target/release/nym-bridge",
+        "/usr/bin/nym-bridge",
+        "755",
+    ],
+    [
+        "target/release/bridge-cfg",
+        "/usr/bin/bridge-cfg",
+        "755",
+    ],
+    [
+        "../../bridges.template.toml",
+        "/usr/share/doc/nym-bridge/bridges.template.toml",
+        "644",
+    ],
+    [
+        "pkg/doc/nym-bridge.1",
+        "/usr/share/man/man1/nym-bridge.1",
+        "644",
+    ],
+]
+systemd-units = { unit-name = "nym-bridge", stop-on-upgrade = true }
+
+# END DEBIAN PACKAGING
+# ------------------------------------------------------------------------------
