+++ crates/nym-bridges-types/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "nym-bridges-types"
+version = "0.1.0"
+edition = "2024"
+
+[features]
+uniffi-bindings = ["dep:uniffi"]
+typescript-bindings = ["dep:ts-rs", "serde"]
+serde = ["dep:serde"]
+
+[dependencies]
+# bip39 = { workspace = true }
+# si-scale = { workspace = true, features = ["lossy-conversions"] }
+# strum.workspace = true
+# strum_macros.workspace = true
+# time = { workspace = true, features = ["serde", "parsing", "formatting"] }
+# tracing = { workspace = true }
+serde = { workspace = true, features = ["derive", "rc"], optional = true }
+# thiserror.workspace = true
+ts-rs = { workspace = true, optional = true }
+# zeroize.workspace = true
+# ipnetwork = { workspace = true, optional = true }
+uniffi = { workspace = true, optional = true }
+
+[build-dependencies]
+uniffi = { workspace = true, features = ["build"] }
+
+[dev-dependencies]
+serde_json = { workspace = true }
