+++ crates/nym-bridges-types/README.md
@@ -1,0 +1,5 @@
+# Nym Bridge Types
+
+This crate is meant to provide bare bones support for the types used by bridges.  This means that
+types related to bridge functionality can be managed here and then re-used, re-imported, and wired
+in across languages, crates, and projects hopefully without adding your own interop types.
