+++ crates/nym-bridges/Cargo.toml
@@ -1,46 +1,44 @@
+
 [package]
-name = "bridge-tools"
+name = "nym-bridges"
+
+authors = ["Nym Technologies SA"]
 version = "0.2.0"
 edition = "2024"
-publish = false
-authors = ["Nym Technologies SA"]
-description = "Client and debugging utilities for Nym transport bridges."
+description = "Pluggable transport library for securing and obfuscating Nym VPN traffic."
 license = "AGPL-3.0-only"
 repository = "https://github.com/nymtech/nym-bridges"
 homepage = "https://nymtech.net"
+readme = "README.md"
+# Whitelist what ships in the published crate. The package root is the repo
+# root, so without this the tarball would also bundle test-env/ (which holds
+# private ed25519 and WireGuard keys) plus tests/, pkg/, Dockerfile and .github.
+include = ["src/**/*.rs", "/README.md", "/LICENSE"]
 
-[[bin]]
-name = "client"
-path = "src/client.rs"
 
-[[bin]]
-name = "client-udp"
-path = "src/client_udp.rs"
 
-[[bin]]
-name = "udp-echo"
-path = "src/udp_echo.rs"
+[lib]
+crate-type = ["cdylib", "lib"]
 
-[[bin]]
-name = "udp-recv"
-path = "src/udp_recv.rs"
+[lints]
+workspace = true
 
-[[bin]]
-name = "udp-sender"
-path = "src/udp_sender.rs"
+[features]
+default = []
+# TODO: features that expose the featues of the types crate
 
 [dependencies]
 anyhow.workspace = true
 base64.workspace = true
 bs58.workspace = true
 bytes.workspace = true
-clap = { version = "4", features = ["derive"] }
 ed25519-dalek = { workspace = true, features = [
-	"pkcs8",
-	"alloc",
-	"pem",
-	"rand_core",
+    "pkcs8",
+    "alloc",
+    "pem",
+    "rand_core",
 ] }
+futures.workspace = true
 getrandom = { workspace = true, default-features = false }
 hex.workspace = true
 lru-slab.workspace = true
@@ -49,40 +47,46 @@ once_cell.workspace = true
 pin-project-lite.workspace = true
 quinn-proto.workspace = true
 quinn.workspace = true
-rand.workspace = true
-rand_chacha.workspace = true
+rand = { workspace = true, features = ["thread_rng", "sys_rng"] }
 rcgen.workspace = true
 ring.workspace = true
 rustc-hash.workspace = true
 rustls = { workspace = true, default-features = false, features = [
-	"std",
-	"ring",
-	"logging",
+    "std",
+    "ring",
+    "logging",
+    "aws_lc_rs",
 ] }
 rustls-pki-types.workspace = true
 serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
 slab.workspace = true
 socket2.workspace = true
 thiserror.workspace = true
 tinyvec = { workspace = true, features = ["alloc"] }
 tokio = { workspace = true, features = ["full"] }
 tokio-rustls = { workspace = true, default-features = false, features = [
-	"ring",
-	"tls12",
-	"logging",
+    "ring",
+    "tls12",
+    "logging",
 ] }
 tokio-util = { workspace = true, features = ["full"] }
 toml.workspace = true
 tracing.workspace = true
-tracing-subscriber = { workspace = true, default-features = false, features = [
-	"env-filter",
-	"fmt",
-	"ansi",
-	"time",
-	"local-time",
-] }
 uuid.workspace = true
 x509-parser = { workspace = true, features = ["verify"] }
 webpki-roots.workspace = true
 
-nym-bridges = { path = "../" }
+nym-bridges-types = { path = "../nym-bridges-types", features = ["serde"]}
+
+
+[dev-dependencies]
+tracing-subscriber = { workspace = true, default-features = false, features = [
+    "env-filter",
+    "fmt",
+    "ansi",
+    "time",
+    "local-time",
+] }
+tempfile.workspace = true
+
