--- nym-bridge/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "nym-bridge"
-version = "0.2.0"
-edition = "2024"
-publish = false
-authors = ["Nym Technologies SA"]
-description = "Server-side transparent forwarder that unwraps obfuscated Nym VPN bridge traffic."
-license = "AGPL-3.0-only"
-repository = "https://github.com/nymtech/nym-bridges"
-homepage = "https://nymtech.net"
-
-[dependencies]
-anyhow.workspace = true
-clap = { workspace = true, features = ["derive"] }
-quinn.workspace = true
-rustls = { workspace = true, default-features = false, features = [
-	"std",
-	"ring",
-	"logging",
-] }
-tokio = { workspace = true, features = ["full"] }
-tokio-rustls = { workspace = true, default-features = false, features = [
-	"ring",
-	"tls12",
-	"logging",
-] }
-tokio-util = { workspace = true, features = ["full"] }
-tracing.workspace = true
-tracing-subscriber = { workspace = true, default-features = false, features = [
-	"env-filter",
-	"fmt",
-	"ansi",
-	"time",
-	"local-time",
-] }
-
-
-nym-bridges = { path = "../" }
-
-nym-bin-common = "1.21.4"
