+++ pkg/README.md
@@ -6,18 +6,23 @@ Build Debian package for the `nym-bridge` as a service.
 
 ### Quick Build
 
-
+```sh
 # From repository root
-`cargo deb`
+cargo deb -p nym-bridge
+```
 
 ### Check Build Artifacts
 
+```sh
+# Pick the newest built nym-bridge package artifact
+DEB="$(ls -1t target/debian/nym-bridge_*_amd64.deb | head -n1)"
+
 # Check Debian best practices
-`lintian target/debian/nym-bridge_0.1.0-2_amd64.deb`
+lintian "$DEB"
 
 # Inspect package contents
-mkdir debdir
-dpkg-deb -R target/debian/nym-bridge_0.1.0-2_amd64.deb debdir
+mkdir -p debdir
+dpkg-deb -R "$DEB" debdir
 tree debdir
 ```
